Moved `FragmentDefinition` to `ast` package, because it feels that it belongs there naturally, instead of its own package.
